+++ package.json
@@ -43,6 +43,7 @@
     "@types/node": "^20.11.25",
     "@types/react": "^18.2.64",
     "@types/react-dom": "^18.2.21",
+    "react-router-dom": "^6.22.3",
     "typescript": "^5.4.2"
   }
 }
